+++ src/test/java/com/epam/reportportal/junit/step/StepReporterFinishTest.java
@@ -27,7 +27,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
 import java.util.List;
-import java.util.concurrent.Executors;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -53,10 +52,7 @@ public class StepReporterFinishTest {
 		TestUtils.mockLaunch(client, null, null, testClassUuid, testMethodUuid);
 		TestUtils.mockNestedSteps(client, testStepUuidOrder);
 		TestUtils.mockBatchLogging(client);
-		ReportPortalListener.setReportPortal(ReportPortal.create(client,
-				TestUtils.standardParameters(),
-				Executors.newSingleThreadExecutor()
-		));
+		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters(), TestUtils.testExecutor()));
 	}
 
 	@Test