+++ src/test/java/com/epam/reportportal/junit/suite/EmptySuiteNoLaunchTest.java
@@ -25,8 +25,6 @@ import com.epam.reportportal.util.test.CommonUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.concurrent.Executors;
-
 import static com.epam.reportportal.junit.utils.TestUtils.PROCESSING_TIMEOUT;
 import static org.mockito.Mockito.*;
 
@@ -41,8 +39,7 @@ public class EmptySuiteNoLaunchTest {
 	public void setupMock() {
 		TestUtils.mockLaunch(client, null, suiteId, classId, methodId);
 		TestUtils.mockBatchLogging(client);
-		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters(),
-				Executors.newSingleThreadExecutor()));
+		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters(), TestUtils.testExecutor()));
 	}
 
 	@Test