+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Client version updated on [5.1.0-RC-12](https://github.com/reportportal/client-java/releases/tag/5.1.0-RC-12)
+### Fixed
+- ItemTree retrieve by test Description
 
 ## [5.1.0-RC-1]
 ### Changed